+++ org.mwc.debrief.scripting/META-INF/MANIFEST.MF
@@ -20,5 +20,6 @@ Import-Package: org.eclipse.ui.plugin,
  org.mwc.cmap.core.DataTypes.Temporal,
  org.mwc.cmap.core.DataTypes.TrackData,
  org.mwc.cmap.core.ui_support,
+ org.mwc.debrief.core,
  org.mwc.debrief.core.editors
 Bundle-Activator: org.mwc.debrief.scripting.ScriptingActivator
