+++ org.mwc.debrief.scripting/src/org/mwc/debrief/scripting/wrappers/Chart.java
@@ -61,7 +61,4 @@ public class Chart
   {
     return new Coastline(inputStream);
   }
-  
-  /*@WrapToScript
-  public*/ 
 }
