+++ org.mwc.debrief.scripting/src/org/mwc/debrief/scripting/wrappers/Core.java
@@ -14,157 +14,15 @@ import org.eclipse.ui.IWorkbenchPage;
 import org.eclipse.ui.IWorkbenchPart;
 import org.eclipse.ui.IWorkbenchWindow;
 import org.eclipse.ui.PlatformUI;
-import org.eclipse.ui.views.contentoutline.IContentOutlinePage;
-import org.mwc.cmap.core.DataTypes.Temporal.ControllableTime;
-import org.mwc.cmap.core.DataTypes.Temporal.TimeControlPreferences;
-import org.mwc.cmap.core.DataTypes.Temporal.TimeManager;
 import org.mwc.cmap.core.DataTypes.Temporal.TimeProvider;
-import org.mwc.cmap.core.DataTypes.TrackData.TrackManager;
 import org.mwc.cmap.core.ui_support.PartMonitor;
 import org.mwc.debrief.core.editors.PlotEditor;
-import org.mwc.debrief.scripting.wrappers.Layers.DLayers;
 
-import MWC.Algorithms.PlainProjection;
 import MWC.GenericData.Duration;
 import MWC.GenericData.HiResDate;
-import MWC.GenericData.WorldArea;
-import MWC.GenericData.WorldLocation;
 
 public class Core
 {
-
-  public static class DEditor
-  {
-    private final PlotEditor _editor;
-
-    public DEditor(final PlotEditor editor)
-    {
-      _editor = editor;
-    }
-
-    public void fitToWindow()
-    {
-      _editor.getChart().rescale();
-      getLayers().fireModified();
-    }
-
-    public WorldArea getArea()
-    {
-      if (_editor != null)
-      {
-        final PlainProjection proj = (PlainProjection) _editor.getAdapter(
-            PlainProjection.class);
-        return proj.getDataArea();
-      }
-      return null;
-    }
-
-    public WorldLocation getCentre()
-    {
-      final WorldArea area = getArea();
-      if (area != null)
-      {
-        return area.getCentre();
-      }
-      else
-      {
-        return null;
-      }
-    }
-
-    public DLayers getLayers()
-    {
-      if (_editor != null)
-      {
-        final MWC.GUI.Layers layers = (MWC.GUI.Layers) _editor.getAdapter(
-            MWC.GUI.Layers.class);
-        if (layers != null)
-        {
-          return new DLayers(layers);
-        }
-      }
-      return null;
-
-    }
-
-    public HiResDate getTime()
-    {
-      final TimeProvider time = (TimeProvider) _editor.getAdapter(
-          TimeProvider.class);
-      return time.getTime();
-    }
-
-    public TimeControlPreferences getTimeControlPreferences()
-    {
-      if (_editor != null)
-      {
-        TimeControlPreferences timeControlPreferences =
-            (TimeControlPreferences) _editor.getAdapter(
-                TimeControlPreferences.class);
-        if (timeControlPreferences != null)
-        {
-          return timeControlPreferences;
-        }
-      }
-      return null;
-    }
-
-    public TimeManager getTimeManager()
-    {
-      if (_editor != null)
-      {
-        TimeManager timeManager = (TimeManager) _editor.getAdapter(
-            ControllableTime.class);
-        if (timeManager != null)
-        {
-          return timeManager;
-        }
-      }
-      return null;
-    }
-
-    public Tote getTote()
-    {
-      if (_editor != null)
-      {
-        TrackManager trackManager = (TrackManager) _editor.getAdapter(
-            TrackManager.class);
-        if (trackManager != null)
-        {
-          return new Tote(trackManager);
-        }
-      }
-      return null;
-    }
-
-    public Outline getOutline()
-    {
-      if (_editor != null)
-      {
-        IContentOutlinePage outline = (IContentOutlinePage) _editor.getAdapter(
-            IContentOutlinePage.class);
-        if (outline != null)
-        {
-          return new Outline(outline);
-        }
-      }
-      return null;
-    }
-
-    public PlainProjection getMap()
-    {
-      if (_editor != null)
-      {
-        PlainProjection map = (PlainProjection) _editor.getAdapter(
-            PlainProjection.class);
-        if (map != null)
-        {
-          return map;
-        }
-      }
-      return null;
-    }
-  }
 
   /*
    * Here is how to provide default value: @ScriptParameter(defaultValue="-1")
@@ -174,22 +32,22 @@ public class Core
     return new HiResDate(date);
   }
 
+  public static Duration createDuration(final int value, final int units)
+  {
+    return new Duration(value, units);
+  }
+
   public static Font createFont(final String fontName, final int style,
       final int size)
   {
     return new Font(fontName, style, size);
   }
 
-  public static Duration createDuration(final int value, final int units)
-  {
-    return new Duration(value, units);
-  }
-
   /**
    * Creates an opaque sRGB color with the specified red, green, and blue values in the range (0 -
    * 255). The actual color used in rendering depends on finding the best match given the color
    * space available for a given output device. Alpha is defaulted to 255.
-   * 
+   *
    * @param red
    *          the red component
    * @param green
@@ -198,12 +56,17 @@ public class Core
    *          the blue component
    * @return
    */
-  public static Color getColor(final int red, final int green, final int blue)
+  public static Color createColor(final int red, final int green, final int blue)
   {
     return new Color(red, green, blue);
   }
+  
+  public static Plot getActivePlot()
+  {
+    return getPlot(null);
+  }
 
-  public static DEditor getEditor(@ScriptParameter(
+  public static Plot getPlot(@ScriptParameter(
       defaultValue = "unset") final String filename)
   {
     final IWorkbench workbench = PlatformUI.getWorkbench();
@@ -229,7 +92,7 @@ public class Core
                 final IEditorPart instance = editor.getEditor(false);
                 if (instance != null)
                 {
-                  return new DEditor((PlotEditor) instance);
+                  return new Plot((PlotEditor) instance);
                 }
               }
             }
@@ -280,14 +143,14 @@ public class Core
       return;
     }
 
-    final DEditor dEditor = getEditor(null);
+    final Plot dEditor = getPlot(null);
     if (dEditor == null)
     {
       System.err.println("Couldn't get editor");
       return;
     }
 
-    final PlotEditor editor = dEditor._editor;
+    final PlotEditor editor = dEditor.getPlot();
     final IWorkbenchWindow window = editor.getSite().getPage()
         .getWorkbenchWindow();
