--- org.mwc.debrief.scripting/src/org/mwc/debrief/scripting/wrappers/Map.java
@@ -1,20 +1,0 @@
-package org.mwc.debrief.scripting.wrappers;
-/*
- *    Debrief - the Open Source Maritime Analysis Application
- *    http://debrief.info
- *
- *    (C) 2000-2018, Deep Blue C Technology Ltd
- *
- *    This library is free software; you can redistribute it and/or
- *    modify it under the terms of the Eclipse Public License v1.0
- *    (http://www.eclipse.org/legal/epl-v10.html)
- *
- *    This library is distributed in the hope that it will be useful,
- *    but WITHOUT ANY WARRANTY; without even the implied warranty of
- *    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. 
- */
-
-public class Map
-{
-  
-}
